refactor(workspace): extract command palette chrome hook
